UIEUS-242-improve-layout-of-sushi-error-messages rename sushi warning file
